refactored switch instead of select in paid oredr status
