+++ src/components/GeneralInfoMenu/GeneralInfoMenu.module.scss
@@ -19,6 +19,13 @@
   line-height: 18.75px;
 }
 
+.generalInfoMenuForm__title_paidStatus {
+  margin-bottom: 28px;
+
+  @include text($font-size: 16px, $font-weight: 600);
+  line-height: 18.75px;
+}
+
 .generalInfoMenuForm__input_wrapper {
   position: relative;
 }
